Only Copy Rebuilt UBT Binaries When They Exist

+++ Scripts/InstallEngineMods.sh
@@ -221,8 +221,12 @@ REBUILD_UBT() {
   eval "$DOTNET" build "./Engine/Source/Programs/AutomationTool/AutomationTool.csproj" -c Development
 
   # Copy the resulting built dlls to the binaries folder
-  cp -r "$UNREAL_ENGINE_PATH/Engine/Source/Programs/UnrealBuildTool/bin/Development/UnrealBuildTool.dll" "$UNREAL_ENGINE_PATH/Engine/Binaries/DotNET/UnrealBuildTool"
-  cp -r "$UNREAL_ENGINE_PATH/Engine/Source/Programs/AutomationTool/bin/Development/AutomationTool.dll" "$UNREAL_ENGINE_PATH/Engine/Binaries/DotNET/AutomationTool"
+  if [ -f "$UNREAL_ENGINE_PATH/Engine/Source/Programs/UnrealBuildTool/bin/Development/UnrealBuildTool.dll" ]; then
+    cp -r "$UNREAL_ENGINE_PATH/Engine/Source/Programs/UnrealBuildTool/bin/Development/UnrealBuildTool.dll" "$UNREAL_ENGINE_PATH/Engine/Binaries/DotNET/UnrealBuildTool"
+  fi
+  if [ -f "$UNREAL_ENGINE_PATH/Engine/Source/Programs/AutomationTool/bin/Development/AutomationTool.dll" ]; then
+    cp -r "$UNREAL_ENGINE_PATH/Engine/Source/Programs/AutomationTool/bin/Development/AutomationTool.dll" "$UNREAL_ENGINE_PATH/Engine/Binaries/DotNET/AutomationTool"
+  fi
 
   echo -e "\nSuccessfully rebuilt UnrealBuildTool with Tempo mods\n"
 }
